#1451 - Bring ImsCwbWriter in-line with other writers
- Modernized additional tests
- Split CollectionReaderAssert into ReaderAssert and WriterAssert parts.
